+++ src/shared/model/index.ts
@@ -3,3 +3,4 @@ export * from "./InputWithLabel";
 export * from "./InputWithFormProps.type";
 export * from "./LogoutModal.type";
 export * from "./ModalData.type";
+export * from "./OcrSearch.type";
